pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "heartbench"
version = "0.1.0"
description = "Christian Framing and Heart-Level Moral Attribution benchmark and evaluation pipeline"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.37.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "jsonlines>=4.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
